COMMERCE-9023 batch-planner-web - refactor Display Context classes
