+++ modules/apps/batch-planner/batch-planner-web/src/main/java/com/liferay/batch/planner/web/internal/display/context/BatchPlannerPlanManagementToolbarDisplayContext.java
@@ -14,9 +14,6 @@
 
 package com.liferay.batch.planner.web.internal.display.context;
 
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemBuilder;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItemListBuilder;
 import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
@@ -24,10 +21,8 @@ import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
-import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -50,35 +45,6 @@ public class BatchPlannerPlanManagementToolbarDisplayContext
 		super(
 			httpServletRequest, liferayPortletRequest, liferayPortletResponse,
 			searchContainer);
-	}
-
-	@Override
-	public List<DropdownItem> getActionDropdownItems() {
-		return DropdownItemList.of(
-			DropdownItemBuilder.putData(
-				"action", "deleteBatchPlannerPlans"
-			).putData(
-				"deleteBatchPlannerPlansURL",
-				PortletURLBuilder.createActionURL(
-					liferayPortletResponse
-				).setActionName(
-					"/batch_planner/delete_batch_planner_plan"
-				).setCMD(
-					Constants.DELETE
-				).setNavigation(
-					getNavigation()
-				).buildString()
-			).setIcon(
-				"trash"
-			).setLabel(
-				LanguageUtil.get(httpServletRequest, "delete")
-			).setQuickAction(
-				true
-			).build());
-	}
-
-	public List<String> getAvailableActions() {
-		return Arrays.asList("deleteBatchPlannerPlans");
 	}
 
 	@Override
@@ -123,6 +89,11 @@ public class BatchPlannerPlanManagementToolbarDisplayContext
 	}
 
 	@Override
+	public Boolean isSelectable() {
+		return false;
+	}
+
+	@Override
 	protected String getNavigation() {
 		return ParamUtil.getString(
 			liferayPortletRequest, getNavigationParam(), "all");
@@ -135,7 +106,7 @@ public class BatchPlannerPlanManagementToolbarDisplayContext
 
 	@Override
 	protected String[] getOrderByKeys() {
-		return new String[] {"name"};
+		return new String[] {"createDate"};
 	}
 
 }
