+++ modules/apps/batch-planner/batch-planner-web/src/main/java/com/liferay/batch/planner/web/internal/display/context/BatchPlannerPlanTemplateManagementToolbarDisplayContext.java
@@ -14,6 +14,9 @@
 
 package com.liferay.batch.planner.web.internal.display.context;
 
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemBuilder;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItemListBuilder;
 import com.liferay.petra.portlet.url.builder.PortletURLBuilder;
@@ -21,8 +24,10 @@ import com.liferay.portal.kernel.dao.search.SearchContainer;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ParamUtil;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -33,10 +38,10 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Matija Petanjek
  */
-public class BatchPlannerLogManagementToolbarDisplayContext
+public class BatchPlannerPlanTemplateManagementToolbarDisplayContext
 	extends BaseSearchContainerManagementToolbarDisplayContext {
 
-	public BatchPlannerLogManagementToolbarDisplayContext(
+	public BatchPlannerPlanTemplateManagementToolbarDisplayContext(
 		HttpServletRequest httpServletRequest,
 		LiferayPortletRequest liferayPortletRequest,
 		LiferayPortletResponse liferayPortletResponse,
@@ -45,6 +50,35 @@ public class BatchPlannerLogManagementToolbarDisplayContext
 		super(
 			httpServletRequest, liferayPortletRequest, liferayPortletResponse,
 			searchContainer);
+	}
+
+	@Override
+	public List<DropdownItem> getActionDropdownItems() {
+		return DropdownItemList.of(
+			DropdownItemBuilder.putData(
+				"action", "deleteBatchPlannerPlanTemplates"
+			).putData(
+				"deleteBatchPlannerPlanTemplatesURL",
+				PortletURLBuilder.createActionURL(
+					liferayPortletResponse
+				).setActionName(
+					"/batch_planner/delete_batch_planner_plan_template"
+				).setCMD(
+					Constants.DELETE
+				).setNavigation(
+					getNavigation()
+				).buildString()
+			).setIcon(
+				"trash"
+			).setLabel(
+				LanguageUtil.get(httpServletRequest, "delete")
+			).setQuickAction(
+				true
+			).build());
+	}
+
+	public List<String> getAvailableActions() {
+		return Arrays.asList("deleteBatchPlannerPlanTemplates");
 	}
 
 	@Override
@@ -89,11 +123,6 @@ public class BatchPlannerLogManagementToolbarDisplayContext
 	}
 
 	@Override
-	public Boolean isSelectable() {
-		return false;
-	}
-
-	@Override
 	protected String getNavigation() {
 		return ParamUtil.getString(
 			liferayPortletRequest, getNavigationParam(), "all");
@@ -106,7 +135,7 @@ public class BatchPlannerLogManagementToolbarDisplayContext
 
 	@Override
 	protected String[] getOrderByKeys() {
-		return new String[] {"createDate"};
+		return new String[] {"name"};
 	}
 
 }
