+++ modules/apps/batch-planner/batch-planner-web/src/main/java/com/liferay/batch/planner/web/internal/portlet/BatchPlannerPortlet.java
@@ -15,7 +15,7 @@
 package com.liferay.batch.planner.web.internal.portlet;
 
 import com.liferay.batch.planner.constants.BatchPlannerPortletKeys;
-import com.liferay.batch.planner.web.internal.display.context.BatchPlannerLogDisplayContext;
+import com.liferay.batch.planner.web.internal.display.context.BatchPlannerPlanDisplayContext;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
 import com.liferay.portal.kernel.util.WebKeys;
 
@@ -60,7 +60,7 @@ public class BatchPlannerPortlet extends MVCPortlet {
 
 		renderRequest.setAttribute(
 			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			new BatchPlannerLogDisplayContext(renderRequest, renderResponse));
+			new BatchPlannerPlanDisplayContext(renderRequest, renderResponse));
 
 		super.render(renderRequest, renderResponse);
 	}
